Tighten isolation
The fileserver (and anything other than codegen and garner) is run in the
default shell, with a garner-only shell being used for the parts we're
testing. This means we no longer need `devShells.barren` and the isolation shell
doesn't GHC or git (or potentially anything else) leak into it.

This also allows us to fold this test into `just github-ci`, rather than adding
a new job to the GH workflow.

+++ test/smoke-test.sh
@@ -5,14 +5,24 @@ shopt -s inherit_errexit
 
 PROJECT_ROOT="$(git rev-parse --show-toplevel)"
 
-codegen
+function run-in-isolation () {
+  nix shell --ignore-environment "$PROJECT_ROOT#garner" -c "$@"
+}
+
+## verify isolation is working
+if run-in-isolation git rev-parse --show-toplevel 2>/dev/null; then
+  echo "Error: Isolation is broken"
+  exit 1
+fi
+
+run-in-isolation codegen
 fs_proc="$("$PROJECT_ROOT/scripts/fileserver" >/dev/null & jobs -p)"
 sleep 2
 trap "kill -s SIGTERM $fs_proc" EXIT
 
 cd examples/haskell
 expected=42
-res="$(garner run haskellExecutable)"
+res="$(run-in-isolation garner run haskellExecutable)"
 if [[ $res == $expected ]]; then
   echo "Smoke test successful"
 else
